Pivotal ID # 182600130: Remove TSV Feature Flag
Re-structure iTests
